Attempt to fix migration errors

+++ database/migrations/2026_01_04_191420_create_ticket_types_table.php
@@ -4,14 +4,16 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
     public function up(): void
     {
-        Schema::createIfNotExist('ticket_types', function (Blueprint $table) {
+        if (Schema::hasTable('ticket_types'))
+            return;
+
+        Schema::create('ticket_types', function (Blueprint $table) {
             $table->id();
             $table->foreignId('event_id')->constrained('events')->onDelete('cascade');
             $table->string('name');
