+++ database/migrations/2026_01_04_191421_create_payments_table.php
@@ -4,14 +4,16 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
     public function up(): void
     {
-        Schema::createIfNotExist('payments', function (Blueprint $table) {
+        if (Schema::hasTable('payments'))
+            return;
+
+        Schema::create('payments', function (Blueprint $table) {
             $table->id();
             $table->foreignId('order_id')->constrained('orders')->onDelete('cascade');
             $table->string('stripe_payment_intent_id')->unique();
